+++ src/xyz/oihana/schema/constants/traits/LogTrait.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits;
+
+trait LogTrait
+{
+    /**
+     * The attribute key of the 'date' property.
+     */
+    public const string DATE = 'date' ;
+
+    /**
+     * The attribute key of the 'level' property.
+     */
+    public const string LEVEL = 'level' ;
+
+    /**
+     * The attribute key of the 'message' property.
+     */
+    public const string MESSAGE = 'message' ;
+
+    /**
+     * The attribute key of the 'time' property.
+     */
+    public const string TIME = 'time' ;
+}