+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
-- Adds ooop\schema (new package)
-- Adds ooop\schema\Log
-- Adds ooop\schema\Pagination
+- Adds xyz\oihana\schema (new package)
+- Adds xyz\oihana\schema\Log
+- Adds xyz\oihana\schema\Pagination
+
 - Adds the org\schema\constants\Schema constant
 
 ## [1.0.0] - 2025-06-17